# packages/sdk/openapi.json
{
  "components": {
    "schemas": {
...
      "BatchDetailResponse": {
        "properties": {
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "failedCount": {
            "title": "Failedcount",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "itemCount": {
            "title": "Itemcount",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/IngestionItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "readyCount": {
            "title": "Readycount",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/IngestionBatchStatus"
          }
        },
        "required": [
          "id",
          "status",
          "itemCount",
          "readyCount",
          "failedCount",
          "createdAt",
          "completedAt",
          "items"
        ],
        "title": "BatchDetailResponse",
        "type": "object"
      },
...
      "CsvLocator": {
        "properties": {
          "rowEnd": {
            "minimum": 1.0,
            "title": "Rowend",
            "type": "integer"
          },
          "rowStart": {
            "minimum": 1.0,
            "title": "Rowstart",
            "type": "integer"
          },
          "type": {
            "const": "csv",
            "default": "csv",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "rowStart",
          "rowEnd"
        ],
        "title": "CsvLocator",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "traceId": {
            "title": "Traceid",
            "type": "string"
          }
        },
        "required": [
          "message",
          "code",
          "traceId"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "IngestionBatchStatus": {
        "enum": [
          "pending",
          "processing",
          "completed",
          "completed_with_errors",
          "failed"
        ],
        "title": "IngestionBatchStatus",
        "type": "string"
      },
...
      "KnowledgeAnswerCitation": {
        "additionalProperties": false,
        "properties": {
          "chunkId": {
            "format": "uuid",
            "title": "Chunkid",
            "type": "string"
          },
          "excerpt": {
            "title": "Excerpt",
            "type": "string"
          },
          "id": {
            "pattern": "^S[1-6]$",
            "title": "Id",
            "type": "string"
          },
          "locator": {
            "discriminator": {
              "mapping": {
                "csv": "#/components/schemas/CsvLocator",
                "docx": "#/components/schemas/DocxLocator",
                "html": "#/components/schemas/HtmlLocator",
                "markdown": "#/components/schemas/TextLocator",
                "pdf": "#/components/schemas/PdfLocator",
                "pptx": "#/components/schemas/PptxLocator",
                "text": "#/components/schemas/TextLocator",
                "xlsx": "#/components/schemas/XlsxLocator"
              },
              "propertyName": "type"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/PdfLocator"
              },
              {
                "$ref": "#/components/schemas/DocxLocator"
              },
              {
                "$ref": "#/components/schemas/PptxLocator"
              },
              {
                "$ref": "#/components/schemas/XlsxLocator"
              },
              {
                "$ref": "#/components/schemas/CsvLocator"
              },
              {
                "$ref": "#/components/schemas/HtmlLocator"
              },
              {
                "$ref": "#/components/schemas/TextLocator"
              }
            ],
            "title": "Locator"
          },
          "mediaType": {
            "title": "Mediatype",
            "type": "string"
          },
          "resourceId": {
            "format": "uuid",
            "title": "Resourceid",
            "type": "string"
          },
          "resourceVersionId": {
            "format": "uuid",
            "title": "Resourceversionid",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "resourceId",
          "resourceVersionId",
          "chunkId",
          "title",
          "mediaType",
          "excerpt",
          "locator",
          "score",
          "id"
        ],
        "title": "KnowledgeAnswerCitation",
        "type": "object"
      },
      "KnowledgeAnswerParagraph": {
        "additionalProperties": false,
        "properties": {
          "citationIds": {
            "items": {
              "type": "string"
            },
            "maxItems": 6,
            "minItems": 1,
            "title": "Citationids",
            "type": "array"
          },
          "text": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "citationIds"
        ],
        "title": "KnowledgeAnswerParagraph",
        "type": "object"
      },
      "KnowledgeAnswerRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "maxLength": 500,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "KnowledgeAnswerRequest",
        "type": "object"
      },
      "KnowledgeAnswerResponse": {
        "additionalProperties": false,
        "properties": {
          "citations": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeAnswerCitation"
            },
            "maxItems": 6,
            "title": "Citations",
            "type": "array"
          },
          "paragraphs": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeAnswerParagraph"
            },
            "maxItems": 8,
            "title": "Paragraphs",
            "type": "array"
          },
          "retrievalMode": {
            "enum": [
              "hybrid",
              "keyword_fallback"
            ],
            "title": "Retrievalmode",
            "type": "string"
          },
          "status": {
            "enum": [
              "answered",
              "insufficient_evidence"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "retrievalMode",
          "paragraphs",
          "citations"
        ],
        "title": "KnowledgeAnswerResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Cairn API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/knowledge/answers": {
      "post": {
        "operationId": "answer_knowledge_api_v1_projects__project_id__knowledge_answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "description": "Session-bound CSRF token returned by login or session restore.",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeAnswerResponse"
                }
              }
            },
            "description": "项目知识生成式回答",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "会话无效",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "请求来源或 CSRF 令牌无效",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "项目不存在或不可访问",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "项目知识在生成期间发生变化",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "搜索请求无效",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "搜索请求过于频繁",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "Retry-After": {
                "description": "当前搜索限流窗口剩余秒数",
                "schema": {
                  "minimum": 1,
                  "type": "integer"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "服务器内部错误",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "数据库、Embedding 或生成式回答服务暂时不可用",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Answer Knowledge",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/projects/{project_id}/knowledge/batches/{batch_id}": {
      "get": {
        "operationId": "get_knowledge_batch_api_v1_projects__project_id__knowledge_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchDetailResponse"
                }
              }
            },
            "description": "知识导入批次详情",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "会话无效",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "项目或知识资源不存在",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "请求参数无效",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "服务器内部错误",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "数据库或对象存储暂时不可用",
            "headers": {
              "Cache-Control": {
                "description": "防止受保护知识响应被浏览器或中间缓存保存",
                "schema": {
                  "const": "private, no-store",
                  "type": "string"
                }
              },
              "X-Request-ID": {
                "description": "请求追踪标识",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Get Knowledge Batch",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}